% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viridis_quo.R
\name{annotate_quo}
\alias{annotate_quo}
\title{Annotate using quo theme font}
\usage{
annotate_quo(
  geom,
  x = NULL,
  y = NULL,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  xend = NULL,
  yend = NULL,
  family = "Lato",
  ...,
  na.rm = FALSE
)
}
\arguments{
\item{geom}{name of geom to use for annotation}

\item{x, y, xmin, ymin, xmax, ymax, xend, yend}{positioning aesthetics -
you must specify at least one of these.}

\item{family}{Font family to use. Defaults to \code{"Lato"}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
}
\description{
A wrapper for \code{\link[ggplot2:annotate]{ggplot2::annotate()}} that sets the font to Lato.
}
\examples{
library(ggplot2)
library(showtext)

font_add_google("Lato", "Lato")
showtext_auto()
# workaround for https://github.com/yixuan/showtext/issues/51
showtext_opts(dpi = 192)

ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  geom_rlm(method = "loess") +
  annotate_quo("label", x = 5.5, y = 30, label = "hwy increases for displ greater than 5.5") +
  theme_quo()
}
