% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{data}
\name{answers}
\alias{answers}
\title{Example Survey Answers}
\format{
A data frame with 9 rows and 3 variables (Q1, Q2, Q3). Q1 and Q2 have answers from a
five-point Likert scale, Q3 has "Yes" or "No" answers.
}
\usage{
answers
}
\description{
Example answers used for \code{\link[=plot_question]{plot_question()}}.
}
\keyword{internal}
