% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differently.R
\name{differently_palettes}
\alias{differently_palettes}
\title{Security Differently Palettes}
\format{
A list of 3 named vectors of colors.
}
\usage{
differently_palettes
}
\description{
Security Differently branded color palettes. Use \code{\link[=pal_differently]{pal_differently()}} to construct palettes of
desired length.
}
\details{
Primary color palette (\code{primary}):
\itemize{
\item \code{blue}: logo, headings, links
\item \code{orange}, \code{lightblue}, \code{lightorange}: graphics, accents
\item \code{darkgray}: body text
\item \code{lightgray}: alternate background, accents
}

Status color palette (\code{status}): \code{success}, \code{warning}, \code{error}

Chart and data visualization palette (\code{plot}): accessible colors for plotting on a white
background.
}
