% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_colors.R
\name{pal_differently}
\alias{pal_differently}
\alias{differently_pal}
\title{Security Differently palette}
\usage{
pal_differently(name = "plot")

differently_pal(name = "plot")
}
\arguments{
\item{name}{Name of palette, one of: \code{primary}, \code{status}, or \code{plot} (the default).}
}
\value{
A vector of hex code colors.
}
\description{
Construct Security Differently branded color palettes.
}
\examples{
library(scales)
show_col(pal_differently()(6))
show_col(pal_differently("primary")(6))
show_col(pal_differently("status")(3))
}
