% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_colors.R
\name{pal_sd}
\alias{pal_sd}
\alias{sd_pal}
\title{Security Differently palette}
\usage{
pal_sd(name = "plot")

sd_pal(name = "plot")
}
\arguments{
\item{name}{Name of palette, one of: \code{primary}, \code{status}, or \code{plot} (the default).}
}
\value{
A vector of hex code colors.
}
\description{
Construct Security Differently branded color palettes.
}
\examples{
library(scales)
show_col(pal_sd()(6))
show_col(pal_sd("primary")(6))
show_col(pal_sd("status")(3))
}
