% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{data}
\name{questions}
\alias{questions}
\title{Example Survey Questions}
\format{
A named vector of question titles of length 3:
\itemize{
\item Q1 = "I enjoy answering survey questions"
\item Q2 = "I use the Likert scale on my own surveys"
\item Q3 = "I would complete future surveys by this author"
}
}
\usage{
questions
}
\description{
Example questions used for \code{\link[=plot_question]{plot_question()}}.
}
\keyword{internal}
