% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_donut.R
\name{ggplot_donut}
\alias{ggplot_donut}
\title{Create a ggplot Donut (Doughnut) Plot}
\usage{
ggplot_donut(data, group = "group", value = "n", hsize = 2)
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[ggplot2:fortify]{fortify()}}. If not specified,
must be supplied in each layer added to the plot.}

\item{group}{Name of column to use for the plot colors}

\item{value}{Name of column to use for the sector size}

\item{hsize}{Size of the donut hole. Defaults to 2, and must be a non-negative value. A value
of 0 creates a pie chart.}
}
\description{
Create a donut plot from a data frame containing groups and corresponding values with
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\details{
Implemented using \code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}}, \code{\link[ggplot2:coord_polar]{ggplot2::coord_polar()}}, and \code{\link[ggplot2:lims]{ggplot2::xlim()}}, adapted
from guides in \href{https://r-graph-gallery.com/128-ring-or-donut-plot.html}{The R Graph Gallery}
and \href{https://r-charts.com/part-whole/donut-chart-ggplot2/}{R CHARTS}.

Some default plot elements are disabled using \code{\link[ggplot2:theme]{ggplot2::theme()}} and \code{\link[ggplot2:element]{ggplot2::element_blank()}}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(ggplot2)

mtcars |>
  count(cyl) |>
  mutate(cyl = as.factor(cyl)) |>
  ggplot_donut(group = "cyl")
}
