% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{jbplot-package}
\alias{jbplot}
\alias{jbplot-package}
\title{jbplot: A Personal Collection of ggplot2 Themes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

My personal collection of ggplot2 themes and styles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jabenninghoff.github.io/jbplot/}
  \item \url{https://github.com/jabenninghoff/jbplot}
  \item Report bugs at \url{https://github.com/jabenninghoff/jbplot/issues}
}

}
\author{
\strong{Maintainer}: John Benninghoff \email{jbenninghoff@mac.com} (\href{https://orcid.org/0000-0002-6230-4742}{ORCID})

}
\keyword{internal}
