% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_question}
\alias{plot_question}
\title{Plot Survey Question}
\usage{
plot_question(question, df, titles, nudge = 1, wrap = 20, themed = TRUE)
}
\arguments{
\item{question}{name of column containing the question answers as a factor with specified levels.}

\item{df}{data frame containing survey answers.}

\item{titles}{named character vector mapping question column names to question titles.}

\item{nudge}{amount of distance to nudge text label (percentage), passed on to
\code{\link[ggplot2:geom_text]{ggplot2::geom_text()}} as \code{nudge_x}.}

\item{wrap}{character width to wrap response labels, passed on to \code{\link[stringr:str_wrap]{stringr::str_wrap()}} as
\code{width}.}

\item{themed}{apply default theme \code{\link[=theme_quo]{theme_quo(minor = TRUE)}} to the plot?}
}
\description{
Plot answers to a survey question using a horizontal bar chart using ggplot2.
}
\examples{
library(purrr)

map(names(questions), plot_question, answers, questions, nudge = 0.25)
}
